issue/89 merge _internal.h content to common_bang.h, add cnrtQueueSync in matmul bang implementation, and fix misc. issues
