+++ src/infiniop/devices/bang/common_bang.h
@@ -2,18 +2,36 @@
 #define __COMMON_BANG_H__
 
 #include "../../../utils.h"
+#include "../pool.h"
+#include "bang_handle.h"
+#include "cnnl.h"
+#include "cnrt.h"
+#include <functional>
 
-// the maximum NRAM memory is 1024 * 768
-#define NRAM_MAX_SIZE (1024 * 256)
-
-#define GDRAM_MAX_SIZE (1024 * 1024 * 1024)
-
-#ifdef __cplusplus
-extern "C" {
-#endif
 #define CHECK_BANG(API) CHECK_INTERNAL(API, CNNL_STATUS_SUCCESS)
-#ifdef __cplusplus
+
+namespace device::bang {
+
+class Handle::Internal {
+    Pool<cnnlHandle_t> cnnl_handles;
+
+    template <typename T>
+    using Fn = std::function<infiniStatus_t(T)>;
+
+public:
+    infiniStatus_t useCnnl(cnrtQueue_t queue, const Fn<cnnlHandle_t> &f) const;
 };
-#endif
+
+cnnlDataType_t getCnnlDtype(infiniDtype_t dt);
+
+// set cnnl tensor descriptor without strides
+infiniStatus_t setCnnlTensor(cnnlTensorDescriptor_t desc,
+                             const InfiniopTensorDescriptor *layout);
+
+// set cnnl tensor descriptor with strides
+infiniStatus_t setCnnlTensorEx(cnnlTensorDescriptor_t desc,
+                               const InfiniopTensorDescriptor *layout);
+
+} // namespace device::bang
 
 #endif // __COMMON_BANG_H__
